# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger DEMO",
    "description": "This is DEMO"
  },
  "servers": [
    { "url": "/api" }
  ],
  "tags": [
    {
      "name": "Person",
      "description": "Everything about your Pets",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    }
  ],
  "paths": {
    "/get_persons": {
      "get": {
        "tags": [
          "Person"
        ],
        "summary": "Get all person",
        "description": "Get all person",
        "operationId": "getallperson",
        "requestBody": {
          "description": "Get all person in database",
          "content": {
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Person not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/get_person/{person_id}": {
      "get": {
        "tags": [
          "Person"
        ],
        "summary": "Get a person",
        "parameters": [{
          "name": "person_id",
          "in": "path",
          "description": "person id to get",
          "required": true,
          "type": "integer",
          "schema" : {
              "$ref": "#/components/schemas/Id"
          }
        }],
        "description": "Get a person",
        "operationId": "getaperson",
        "requestBody": {
          "description": "Get person with person id",
          "content": {
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Person not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Person": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "lastname": {
            "type": "string",
            "example": "theUser"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "age": {
            "type": "integer",
            "example": "21"
          }
        },
        "xml": {
          "name": "person"
        }
      },
      "Id": {
        "type": "integer",
        "format":"int64"
      }
    }
  }
}